refactor(system): extract game-path detection + cache cleanup into a service
commands/system.rs drops from 1044 to 825 lines. Moved the registry/fs
game-environment side effects into new services/game_env_service.rs:

- detect_game_path (+ extract_ini_value helper): beanfun INI + HKCU/HKLM
  registry lookup for the MapleStory install path
- cleanup_game_cache: remove cache dirs, *.$$$ leftovers, *.dmp dumps, stale DLLs

Command handlers (detect_game_path, detect_game_path_inner, cleanup_game_cache)
now delegate; names/signatures unchanged so lib.rs + frontend are untouched.
Pure mechanical move.
